django_deployment/requirements.txt
pandas
numpy
django
djangorestframework
scikit-learn
joblib
